requirements.txt
# MCP server
fastmcp>=3.0,<4.0

# RAG pipeline
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
chromadb>=0.6,<1.0
sentence-transformers>=3.0,<4.0

# Environment
python-dotenv>=1.2,<2.0

# SSH transport
scrapli2>=2.0.0rc1

# NetBox inventory
pynetbox>=7.0,<8.0

# HashiCorp Vault
hvac>=2.0,<3.0
